shortener/shortener_service.go
package shortener

import (
	"crypto/sha256"
	"fmt"
	"math/big"

	"github.com/itchyny/base58-go"
)

func createSha256(input string) []byte {
	a := sha256.New()
	a.Write([]byte(input))
	return a.Sum(nil)
}

func base58Encoded(bytes []byte) (string, error) {
	encoding := base58.BitcoinEncoding
	encoded, err := encoding.Encode(bytes)
	if err != nil {
		fmt.Println(err.Error())
		return "", err
	}
	return string(encoded), nil
}

func GenerateShortLink(initialLink string) (string, error) {
	urlHashBytes := createSha256(initialLink)
	generatedNumber := new(big.Int).SetBytes(urlHashBytes).Uint64()
	finalString, err := base58Encoded([]byte(fmt.Sprintf("%d", generatedNumber)))
	if err != nil {
		return "", err
	}
	return finalString[:6], nil
}
